+++ packages/deps-analysis/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @t-care/deps-analysis
 
+## 1.3.1
+
+### Patch Changes
+
+- fix cache error
+- Updated dependencies
+  - @t-care/utils@1.2.7
+
+## 1.3.1-beta.0
+
+### Patch Changes
+
+- fix cache error
+- Updated dependencies
+  - @t-care/utils@1.2.7-beta.0
+
 ## 1.3.0
 
 ### Minor Changes